+++ src/apis/party.ts
@@ -1,9 +1,4 @@
-import type {
-  PartyApplicationData,
-  PartyUserListByAdminResponse,
-  PartyUserResponse,
-  SinglePartyResponse,
-} from '@/types/party';
+import type { PartyUserListByAdminResponse, PartyUserResponse, SinglePartyResponse } from '@/types/party';
 
 import { fileUploadApi, privateApi, publicApi } from '.';
 
@@ -108,59 +103,59 @@ export const fetchPatchPartyInfo = async ({
 };
 
 // 파티 모집 생성하기
-export const fetchPostRecruitmentParty = async ({
-  partyId,
-  positionId,
-  content,
-  recruiting_count,
-}: {
-  partyId: number;
-  positionId: number;
-  content: string;
-  recruiting_count: number;
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/recruitments`, {
-      positionId,
-      content,
-      recruiting_count,
-    });
-    return response.data;
-  } catch (error) {
-    console.error('fetchPostRecruitmentParty error:', error);
-    return error;
-  }
-};
+// export const fetchPostRecruitmentParty = async ({
+//   partyId,
+//   positionId,
+//   content,
+//   recruiting_count,
+// }: {
+//   partyId: number;
+//   positionId: number;
+//   content: string;
+//   recruiting_count: number;
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/recruitments`, {
+//       positionId,
+//       content,
+//       recruiting_count,
+//     });
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchPostRecruitmentParty error:', error);
+//     return error;
+//   }
+// };
 
-// 파티 지원하기 페이지
-// API 응답 타입 정의
-export interface ApplyPartyResponse {
-  id: number;
-  message: string;
-  status: 'pending' | 'processing' | 'approved' | 'rejected'; // 상태값 제한
-  createdAt: string; // ISO 8601 형식의 날짜 문자열
-}
+// // 파티 지원하기 페이지
+// // API 응답 타입 정의
+// export interface ApplyPartyResponse {
+//   id: number;
+//   message: string;
+//   status: 'pending' | 'processing' | 'approved' | 'rejected'; // 상태값 제한
+//   createdAt: string; // ISO 8601 형식의 날짜 문자열
+// }
 
-export const fetchPostApplyParty = async ({
-  partyId,
-  partyRecruitmentId,
-  body,
-}: {
-  partyId: number;
-  partyRecruitmentId: number;
-  body: { message: string };
-}): Promise<ApplyPartyResponse> => {
-  try {
-    const response = await privateApi.post<ApplyPartyResponse>(
-      `/parties/${partyId}/recruitments/${partyRecruitmentId}/applications`,
-      body,
-    );
-    return response.data;
-  } catch (error) {
-    console.error('fetchPostApplyParty error : ', error);
-    throw error;
-  }
-};
+// export const fetchPostApplyParty = async ({
+//   partyId,
+//   partyRecruitmentId,
+//   body,
+// }: {
+//   partyId: number;
+//   partyRecruitmentId: number;
+//   body: { message: string };
+// }): Promise<ApplyPartyResponse> => {
+//   try {
+//     const response = await privateApi.post<ApplyPartyResponse>(
+//       `/parties/${partyId}/recruitments/${partyRecruitmentId}/applications`,
+//       body,
+//     );
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchPostApplyParty error : ', error);
+//     throw error;
+//   }
+// };
 
 // 파티 페이지 - 홈탭
 export const fetchGetPartyHome = async ({ partyId }: { partyId: number }) => {
@@ -226,7 +221,7 @@ export const fetchPostReports = async ({
   content: string;
 }) => {
   try {
-    const response = await privateApi.post(`reports`, {
+    const response = await privateApi.post(`/reports`, {
       type,
       typeId,
       content,
@@ -237,38 +232,38 @@ export const fetchPostReports = async ({
   }
 };
 
-// 파티 모집 목록 조회
-/**
- * @Params partyId(파티 아이디; number), sort('createdAt'), order('ASC', 'DESC'), status('active', 'completed'), main('기획자', '디자이너', '개발자', '마케터/광고')
- */
-export const fetchGetPartyRecruitmentsList = async ({
-  partyId,
-  sort = 'createdAt',
-  order = 'ASC',
-  status = 'active',
-  main,
-}: {
-  partyId: number;
-  status: string;
-  sort?: string;
-  order?: string;
-  main?: string;
-}) => {
-  try {
-    const response = await privateApi.get(`/parties/${partyId}/recruitments`, {
-      params: {
-        sort,
-        order,
-        status,
-        ...(main && { main }), // main 값이 있을 때만 쿼리 파라미터에 포함
-      },
-    });
-    return response.data;
-  } catch (error) {
-    console.error('fetchGetPartyRecruitmentsList error: ', error);
-    throw error;
-  }
-};
+// // 파티 모집 목록 조회
+// /**
+//  * @Params partyId(파티 아이디; number), sort('createdAt'), order('ASC', 'DESC'), status('active', 'completed'), main('기획자', '디자이너', '개발자', '마케터/광고')
+//  */
+// export const fetchGetPartyRecruitmentsList = async ({
+//   partyId,
+//   sort = 'createdAt',
+//   order = 'ASC',
+//   status = 'active',
+//   main,
+// }: {
+//   partyId: number;
+//   status: string;
+//   sort?: string;
+//   order?: string;
+//   main?: string;
+// }) => {
+//   try {
+//     const response = await privateApi.get(`/parties/${partyId}/recruitments`, {
+//       params: {
+//         sort,
+//         order,
+//         status,
+//         ...(main && { main }), // main 값이 있을 때만 쿼리 파라미터에 포함
+//       },
+//     });
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchGetPartyRecruitmentsList error: ', error);
+//     throw error;
+//   }
+// };
 
 // 파티 모집 상세 조회
 export const fetchGetPartyRecruitments = async ({ partyRecruitmentId }: { partyRecruitmentId: number }) => {
@@ -281,115 +276,115 @@ export const fetchGetPartyRecruitments = async ({ partyRecruitmentId }: { partyR
 };
 
 // 여러 개의 특정 파티 모집 삭제
-export const fetchDeletePartyRecruitments = async ({
-  partyId,
-  recruitmentIds,
-}: {
-  partyId: number;
-  recruitmentIds: number[];
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/admin/recruitments/batch-delete`, {
-      partyRecruitmentIds: recruitmentIds,
-    });
-    return response.data;
-  } catch (error) {
-    console.error('fetchDeletePartyRecruitments error:', error);
-  }
-};
+// export const fetchDeletePartyRecruitments = async ({
+//   partyId,
+//   recruitmentIds,
+// }: {
+//   partyId: number;
+//   recruitmentIds: number[];
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/admin/recruitments/batch-delete`, {
+//       partyRecruitmentIds: recruitmentIds,
+//     });
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchDeletePartyRecruitments error:', error);
+//   }
+// };
 
 // 하나의 특정 파티 모집 삭제
-export const fetchDeletePartyRecruitmentOnly = async ({
-  partyId,
-  partyRecruitmentId,
-}: {
-  partyId: number;
-  partyRecruitmentId: number;
-}) => {
-  try {
-    const response = await privateApi.delete(`/parties/${partyId}/admin/recruitments/${partyRecruitmentId}`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchDeletePartyRecruitmentOnly error:', error);
-    throw error;
-  }
-};
+// export const fetchDeletePartyRecruitmentOnly = async ({
+//   partyId,
+//   partyRecruitmentId,
+// }: {
+//   partyId: number;
+//   partyRecruitmentId: number;
+// }) => {
+//   try {
+//     const response = await privateApi.delete(`/parties/${partyId}/admin/recruitments/${partyRecruitmentId}`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchDeletePartyRecruitmentOnly error:', error);
+//     throw error;
+//   }
+// };
 
 // 파티 모집 수정
-export const fetchUpdatePartyRecruitment = async ({
-  partyId,
-  partyRecruitmentId,
-  positionId,
-  content,
-  recruiting_count,
-}: {
-  partyId: number;
-  partyRecruitmentId: number;
-  positionId: number;
-  content: string;
-  recruiting_count: number;
-}) => {
-  try {
-    const response = await privateApi.patch(`/parties/${partyId}/recruitments/${partyRecruitmentId}`, {
-      positionId,
-      content,
-      recruiting_count,
-    });
-    return response.data;
-  } catch (error) {
-    console.error('fetchUpdatePartyRecruitment error:', error);
-  }
-};
+// export const fetchUpdatePartyRecruitment = async ({
+//   partyId,
+//   partyRecruitmentId,
+//   positionId,
+//   content,
+//   recruiting_count,
+// }: {
+//   partyId: number;
+//   partyRecruitmentId: number;
+//   positionId: number;
+//   content: string;
+//   recruiting_count: number;
+// }) => {
+//   try {
+//     const response = await privateApi.patch(`/parties/${partyId}/recruitments/${partyRecruitmentId}`, {
+//       positionId,
+//       content,
+//       recruiting_count,
+//     });
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchUpdatePartyRecruitment error:', error);
+//   }
+// };
 
-// 파티 모집 완료 처리
-export const fetchCompletePartyRecruitment = async ({
-  partyId,
-  partyRecruitmentId,
-}: {
-  partyId: number;
-  partyRecruitmentId: number;
-}) => {
-  try {
-    const response = await privateApi.patch(`/parties/${partyId}/admin/recruitment/${partyRecruitmentId}/completed`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchCompletePartyRecruitment error:', error);
-    throw error;
-  }
-};
+// // 파티 모집 완료 처리
+// export const fetchCompletePartyRecruitment = async ({
+//   partyId,
+//   partyRecruitmentId,
+// }: {
+//   partyId: number;
+//   partyRecruitmentId: number;
+// }) => {
+//   try {
+//     const response = await privateApi.patch(`/parties/${partyId}/admin/recruitment/${partyRecruitmentId}/completed`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchCompletePartyRecruitment error:', error);
+//     throw error;
+//   }
+// };
 
 // 파티 모집 여러개 완료 처리
-export const fetchCompletePartyRecruitmentBatchUpdate = async ({
-  partyId,
-  partyRecruitmentIds,
-}: {
-  partyId: number;
-  partyRecruitmentIds: number[];
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/admin/recruitment/batch-status`, {
-      partyRecruitmentIds,
-    });
+// export const fetchCompletePartyRecruitmentBatchUpdate = async ({
+//   partyId,
+//   partyRecruitmentIds,
+// }: {
+//   partyId: number;
+//   partyRecruitmentIds: number[];
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/admin/recruitment/batch-status`, {
+//       partyRecruitmentIds,
+//     });
 
-    if (response.status === 204) {
-      return { success: true };
-    }
-  } catch (error) {
-    console.error('fetchCompletePartyRecruitmentBatchUpdate error:', error);
-    return { success: false, error };
-  }
-};
+//     if (response.status === 204) {
+//       return { success: true };
+//     }
+//   } catch (error) {
+//     console.error('fetchCompletePartyRecruitmentBatchUpdate error:', error);
+//     return { success: false, error };
+//   }
+// };
 
 // 파티 모집 단일 조회 /dev/api/parties/recruitments/{partyRecruitmentId}
-export const fetchPartyRecruitmentDetails = async (partyRecruitmentId: number) => {
-  try {
-    const response = await privateApi.get(`/parties/recruitments/${partyRecruitmentId}`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchPartyRecruitmentDetails error:', error);
-    return null;
-  }
-};
+// export const fetchPartyRecruitmentDetails = async (partyRecruitmentId: number) => {
+//   try {
+//     const response = await privateApi.get(`/parties/recruitments/${partyRecruitmentId}`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchPartyRecruitmentDetails error:', error);
+//     return null;
+//   }
+// };
 
 // [GET] 관리자-파티원 목록 조회
 export const fetchPartyAdminUsers = async ({
@@ -483,122 +478,122 @@ export const fetchUpdatePartyUserPosition = async ({
   }
 };
 
-// [GET] 파티 포지션 모집별 지원자 목록 조회 API
-export const fetchPartyRecruitmentApplications = async ({
-  partyId,
-  partyRecruitmentId,
-  page = 1,
-  limit = 5,
-  sort = 'createdAt',
-  order = 'ASC',
-  status,
-}: {
-  partyId: number;
-  partyRecruitmentId: number;
-  page?: number;
-  limit?: number;
-  sort?: 'createdAt';
-  order?: string;
-  status?: 'processing' | 'approved' | 'pending' | 'rejected';
-}): Promise<PartyApplicationData> => {
-  try {
-    const response = await privateApi.get(`/parties/${partyId}/recruitments/${partyRecruitmentId}/applications`, {
-      params: {
-        page,
-        limit,
-        sort,
-        order,
-        ...(status && { status }), // status 값이 있을 때만 포함
-      },
-    });
-    return response.data;
-  } catch (error) {
-    console.error('fetchPartyRecruitmentApplications error:', error);
-    throw error;
-  }
-};
+// // [GET] 파티 포지션 모집별 지원자 목록 조회 API
+// export const fetchPartyRecruitmentApplications = async ({
+//   partyId,
+//   partyRecruitmentId,
+//   page = 1,
+//   limit = 5,
+//   sort = 'createdAt',
+//   order = 'ASC',
+//   status,
+// }: {
+//   partyId: number;
+//   partyRecruitmentId: number;
+//   page?: number;
+//   limit?: number;
+//   sort?: 'createdAt';
+//   order?: string;
+//   status?: 'processing' | 'approved' | 'pending' | 'rejected';
+// }): Promise<PartyApplicationData> => {
+//   try {
+//     const response = await privateApi.get(`/parties/${partyId}/recruitments/${partyRecruitmentId}/applications`, {
+//       params: {
+//         page,
+//         limit,
+//         sort,
+//         order,
+//         ...(status && { status }), // status 값이 있을 때만 포함
+//       },
+//     });
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchPartyRecruitmentApplications error:', error);
+//     throw error;
+//   }
+// };
 
-// [POST] (지원자) 파티 지원자 승인 /dev/api/parties/{partyId}/applications/{partyApplicationId}/approval
-export const fetchApprovePartyApplication = async ({
-  partyId,
-  partyApplicationId,
-}: {
-  partyId: number;
-  partyApplicationId: number;
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/applications/${partyApplicationId}/approval`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchApprovePartyApplication error:', error);
-    return error;
-  }
-};
+// // [POST] (지원자) 파티 지원자 승인 /dev/api/parties/{partyId}/applications/{partyApplicationId}/approval
+// export const fetchApprovePartyApplication = async ({
+//   partyId,
+//   partyApplicationId,
+// }: {
+//   partyId: number;
+//   partyApplicationId: number;
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/applications/${partyApplicationId}/approval`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchApprovePartyApplication error:', error);
+//     return error;
+//   }
+// };
 
 // [POST] (파티장) 파티 지원자 승인 /dev/api/parties/{partyId}/applications/{partyApplicationId}/approval
-export const fetchAdminApprovePartyApplication = async ({
-  partyId,
-  partyApplicationId,
-}: {
-  partyId: number;
-  partyApplicationId: number;
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/admin/applications/${partyApplicationId}/approval`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchApprovePartyApplication error:', error);
-    return error;
-  }
-};
+// export const fetchAdminApprovePartyApplication = async ({
+//   partyId,
+//   partyApplicationId,
+// }: {
+//   partyId: number;
+//   partyApplicationId: number;
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/admin/applications/${partyApplicationId}/approval`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchApprovePartyApplication error:', error);
+//     return error;
+//   }
+// };
 
-// [POST] (지원자) 파티 지원자 거절 /dev/api/parties/{partyId}/applications/{partyApplicationId}/rejection
-export const fetchRejectPartyApplication = async ({
-  partyId,
-  partyApplicationId,
-}: {
-  partyId: number;
-  partyApplicationId: number;
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/applications/${partyApplicationId}/rejection`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchRejectPartyApplication error:', error);
-    return error;
-  }
-};
+// // [POST] (지원자) 파티 지원자 거절 /dev/api/parties/{partyId}/applications/{partyApplicationId}/rejection
+// export const fetchRejectPartyApplication = async ({
+//   partyId,
+//   partyApplicationId,
+// }: {
+//   partyId: number;
+//   partyApplicationId: number;
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/applications/${partyApplicationId}/rejection`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchRejectPartyApplication error:', error);
+//     return error;
+//   }
+// };
 
 // [POST] (파티장) 파티 지원자 거절 /dev/api/parties/{partyId}/applications/{partyApplicationId}/rejection
-export const fetchAdminRejectPartyApplication = async ({
-  partyId,
-  partyApplicationId,
-}: {
-  partyId: number;
-  partyApplicationId: number;
-}) => {
-  try {
-    const response = await privateApi.post(`/parties/${partyId}/admin/applications/${partyApplicationId}/rejection`);
-    return response.data;
-  } catch (error) {
-    console.error('fetchRejectPartyApplication error:', error);
-    return error;
-  }
-};
+// export const fetchAdminRejectPartyApplication = async ({
+//   partyId,
+//   partyApplicationId,
+// }: {
+//   partyId: number;
+//   partyApplicationId: number;
+// }) => {
+//   try {
+//     const response = await privateApi.post(`/parties/${partyId}/admin/applications/${partyApplicationId}/rejection`);
+//     return response.data;
+//   } catch (error) {
+//     console.error('fetchRejectPartyApplication error:', error);
+//     return error;
+//   }
+// };
 
-// [DELETE] (지원자) 파티 지원 삭제(취소)
-export const fetchDeletePartyApplication = async ({
-  partyId,
-  partyApplicationId,
-}: {
-  partyId: number;
-  partyApplicationId: number;
-}): Promise<void> => {
-  try {
-    const response = await privateApi.delete(`/parties/${partyId}/applications/${partyApplicationId}`);
-    console.log('파티 지원 삭제 성공:', response.status);
-  } catch (error) {
-    console.error('fetchDeletePartyApplication error:', error);
-    throw error;
-  }
-};
+// // [DELETE] (지원자) 파티 지원 삭제(취소)
+// export const fetchDeletePartyApplication = async ({
+//   partyId,
+//   partyApplicationId,
+// }: {
+//   partyId: number;
+//   partyApplicationId: number;
+// }): Promise<void> => {
+//   try {
+//     const response = await privateApi.delete(`/parties/${partyId}/applications/${partyApplicationId}`);
+//     console.log('파티 지원 삭제 성공:', response.status);
+//   } catch (error) {
+//     console.error('fetchDeletePartyApplication error:', error);
+//     throw error;
+//   }
+// };
